OK SO IT TURNS OUT JAVAFX SUCKS!!!
i have been informed by Stack Overflow that you can't actually do a fat
jar with javafx and have it be platform-independent, in addition to all
the javafx/java 12/tornadofx dependencies being a pit of misery

argh. well. now I search for a library that I can draw things on and
make build. this shall be an Adventure.